NO-JIRA Fix alloc-size-larger-than warning from gcc 7
The following warning is fixed by this commit

```
/home/jdanek/Work/repos/qpid-dispatch/src/entity.c: In function ‘qd_entity_set_stringf’:
/home/jdanek/Work/repos/qpid-dispatch/src/entity.c:235:10: error: argument 1 range [18446744071562067968, 18446744073709551615] exceeds maximum object size 9223372036854775807 [-Werror=alloc-size-larger-than=]
     char buf[len+1];
          ^~~
/home/jdanek/Work/repos/qpid-dispatch/src/entity.c:235:10: note: in a call to built-in allocation function ‘__builtin_alloca_with_align’
/home/jdanek/Work/repos/qpid-dispatch/src/entity.c:235:10: error: argument 1 range [18446744071562067968, 18446744073709551615] exceeds maximum object size 9223372036854775807 [-Werror=alloc-size-larger-than=]
/home/jdanek/Work/repos/qpid-dispatch/src/entity.c:235:10: note: in a call to built-in allocation function ‘__builtin_alloca_with_align’
cc1: all warnings being treated as errors
make[2]: *** [src/CMakeFiles/qpid-dispatch.dir/build.make:273: src/CMakeFiles/qpid-dispatch.dir/entity.c.o] Error 1
make[2]: *** Waiting for unfinished jobs....
make[1]: *** [CMakeFiles/Makefile2:982: src/CMakeFiles/qpid-dispatch.dir/all] Error 2
make: *** [Makefile:141: all] Error 2
```

+++ src/entity.c
@@ -229,7 +229,7 @@ qd_error_t qd_entity_set_stringf(qd_entity_t *entity, const char* attribute, con
     char dummy[1];
     va_list ap;
     va_start(ap, format);
-    int len = vsnprintf(dummy, 1, format, ap);
+    size_t len = vsnprintf(dummy, 1, format, ap);
     va_end(ap);
 
     char buf[len+1];
